Early prototype has released
In this version, the application can download file from input url and then read file contents.
All the contents will be normalized and imported into database for future analysis.
developed analyze method in conventional approach which is mainly relies on database query and has a good performance in import and analyze
the result will be saved into database table
there are a bunch of UI improvement which will be applied to application very soon